move for all function using the interaction classifier in the interaction directory
